feat: refactor conditional formatting utilities and improve code organization
