+++ webapp/packages/plugin-data-viewer-conditional-formatting/src/utils/clamp.ts
@@ -1,0 +1,11 @@
+/*
+ * CloudBeaver - Cloud Database Manager
+ * Copyright (C) 2020-2025 DBeaver Corp and others
+ *
+ * Licensed under the Apache License, Version 2.0.
+ * you may not use this file except in compliance with the License.
+ */
+
+export function clamp(x: number, a: number, b: number): number {
+  return Math.max(a, Math.min(b, x));
+}
