+++ webapp/packages/plugin-data-viewer-conditional-formatting/src/utils/getColorMix.ts
@@ -6,11 +6,9 @@
  * you may not use this file except in compliance with the License.
  */
 
+import { clamp } from './clamp.js';
+
 export function getColorMix(color1: string, color2: string, ratio: number): string {
   const percent = clamp(ratio * 100, 0, 100);
   return `color-mix(in oklab, ${color1} calc(100% - ${percent}%), ${color2} ${percent}%)`;
-}
-
-export function clamp(x: number, a: number, b: number): number {
-  return Math.max(a, Math.min(b, x));
 }
